+++ tests/manual_bnb_chain_check_revert_reason.py
@@ -13,7 +13,7 @@ from web3.datastructures import AttributeDict
 
 from eth_defi.abi import get_deployed_contract
 from eth_defi.gas import node_default_gas_price_strategy
-from eth_defi.txmonitor import wait_transactions_to_complete
+from eth_defi.confirmation import wait_transactions_to_complete
 from web3.middleware import construct_sign_and_send_raw_middleware